requirements.txt
certifi==2025.4.26
charset-normalizer==2.0.12
idna==3.10
influxdb==5.3.2
msgpack==1.0.5
paho-mqtt==1.6.1
pkg_resources==0.0.0
python-dateutil==2.9.0.post0
pytz==2025.2
requests==2.27.1
six==1.17.0
urllib3==1.26.20
